LayoutTests/imported/w3c/web-platform-tests/css/css-align/parsing/place-self-computed-expected.txt


FAIL Property place-self value 'auto auto' assert_equals: expected "auto" but got "auto auto"
FAIL Property place-self value 'normal' assert_equals: expected "normal" but got "normal normal"
FAIL Property place-self value 'stretch' assert_equals: expected "stretch" but got "stretch stretch"
FAIL Property place-self value 'first baseline' assert_equals: expected "baseline" but got "baseline baseline"
FAIL Property place-self value 'last baseline last baseline' assert_equals: expected "last baseline" but got "last baseline last baseline"
FAIL Property place-self value 'center center' assert_equals: expected "center" but got "center center"
FAIL Property place-self value 'start' assert_equals: expected "start" but got "start start"
FAIL Property place-self value 'self-start' assert_equals: expected "self-start" but got "self-start self-start"
FAIL Property place-self value 'flex-end' assert_equals: expected "flex-end" but got "flex-end flex-end"
FAIL Property place-self value 'unsafe center' assert_equals: expected "unsafe center" but got "unsafe center unsafe center"
FAIL Property place-self value 'safe self-end safe self-end' assert_equals: expected "safe self-end" but got "safe self-end safe self-end"
PASS Property place-self value 'auto last baseline'
PASS Property place-self value 'baseline flex-end'
PASS Property place-self value 'unsafe center stretch'
PASS Property place-self value 'normal right'
PASS Property place-self value 'baseline unsafe left'
